add look and menu panels
