+++ crates/client/src/main.rs
@@ -55,7 +55,7 @@ use crate::{
         panels::{
             get_player_inventory, render_body_panel, render_crafting_panel, render_credits_panel,
             render_help_panel, render_inventory_list_only, render_inventory_panel,
-            render_quit_panel,
+            render_look_panel, render_menu_panel, render_quit_panel,
         },
     },
 };
@@ -1572,166 +1572,4 @@ fn render_bottom_menu(f: &mut Frame, app: &mut App, area: Rect) {
     f.render_widget(tabs, area);
 }
 
-fn render_menu_panel(f: &mut Frame, app: &mut App, area: Rect) {
-    let mut lines: Vec<Line> = Vec::new();
-    lines.push(Line::from(Span::styled(
-        "Game Menu",
-        Style::default().fg(Color::Cyan),
-    )));
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::raw(
-        "Use configured keybinds to Save/Load/Quit (see config).",
-    )));
-    lines.push(Line::from(""));
-    // Config source
-    lines.push(Line::from(Span::styled(
-        "Configuration:",
-        Style::default().fg(Color::Cyan),
-    )));
-    lines.push(Line::from(Span::raw(format!(
-        "  {}",
-        app.game.res.config.source_label()
-    ))));
-    lines.push(Line::from(Span::raw(format!(
-        "  Player: {}",
-        app.game.res.player_name
-    ))));
-    lines.push(Line::from(Span::raw(format!(
-        "  Developer mode: {}",
-        if app.game.res.developer_mode {
-            "ON"
-        } else {
-            "OFF"
-        }
-    ))));
-    lines.push(Line::from(""));
-    // Logging location information
-    lines.push(Line::from(Span::raw("Logging:")));
-    lines.push(Line::from(Span::raw(format!(
-        "  Path: {}",
-        app.log_full_path
-    ))));
-    lines.push(Line::from(""));
-    lines.push(Line::from(Span::raw(format!(
-        "Seed: {}",
-        app.game.res.seed
-    ))));
-    lines.push(Line::from(Span::raw(format!(
-        "Tick: {}",
-        app.game.res.gametick
-    ))));
-    lines.push(Line::from(Span::raw(format!(
-        "View Z: {}",
-        app.game.res.view_z
-    ))));
-    if let Some(e) = app.game.res.player_entity {
-        if let Ok(pos) = app
-            .game
-            .world
-            .get::<&lithicrivers_core::components::Position>(e)
-        {
-            lines.push(Line::from(Span::raw(format!(
-                "Player: ({}, {}, {})",
-                pos.x, pos.y, pos.z
-            ))));
-        }
-    }
-    let block = Block::default().borders(Borders::ALL).title("Menu");
-    let inner = block.inner(area);
-    let p = Paragraph::new(lines).alignment(Alignment::Left);
-    f.render_widget(p, inner);
-    f.render_widget(block, area);
-}
-
 // Helper function to get player's inventory as a HashMap
-
-fn render_look_panel(f: &mut Frame, app: &mut App, area: Rect) {
-    let pos = app.look_cursor;
-
-    let mut lines: Vec<Line> = Vec::new();
-
-    lines.push(Line::from(Span::raw(format!(
-        "Pos: ({}, {}, {})",
-        pos.x, pos.y, pos.z
-    ))));
-
-    lines.push(Line::from(Span::raw(format!(
-        "Chunk: ({}, {}, {})",
-        pos.x / CHUNK_SIZE,
-        pos.y / CHUNK_SIZE,
-        pos.z
-    ))));
-
-    lines.push(Line::from(Span::raw(format!(""))));
-
-    // Entity and item info
-    let mut any_entity = false;
-    for (_e, (e_pos, maybe_player, maybe_sr, maybe_drop)) in app
-        .game
-        .world
-        .query::<(
-            &lithicrivers_core::components::Position,
-            Option<&lithicrivers_core::components::Player>,
-            Option<&lithicrivers_core::components::SpriteRef>,
-            Option<&lithicrivers_core::components::DroppedItem>,
-        )>()
-        .iter()
-    {
-        if *e_pos == pos {
-            any_entity = true;
-            if maybe_player.is_some() {
-                lines.push(Line::from(Span::styled(
-                    "Entity: Player",
-                    Style::default().fg(Color::Green),
-                )));
-            } else if let Some(sr) = maybe_sr {
-                lines.push(Line::from(Span::raw(format!(
-                    "Entity: {}::{}",
-                    sr.category, sr.name
-                ))));
-            } else if let Some(di) = maybe_drop {
-                lines.push(Line::from(Span::raw(format!(
-                    "Dropped: {:?} x{}",
-                    di.kind, di.qty
-                ))));
-            } else {
-                lines.push(Line::from(Span::raw("Entity: (unknown)")));
-            }
-        }
-    }
-    if !any_entity {
-        lines.push(Line::from(Span::raw("Entities: (none)")));
-    }
-
-    lines.push(Line::from("Entity art:"));
-    // Gather 12x8 art for entity under cursor (if any)
-    if entity_art_12x8_lines_for_position(app, pos, &mut lines) {
-        lines.push(Line::from(""));
-    } else {
-        lines.extend(empty_art_12x8_lines_for_position());
-        lines.push(Line::from(""));
-    }
-
-    // Tile info
-    let tile_kind = app.game.res.world.get_tile(pos.x, pos.y, pos.z);
-    lines.push(Line::from(Span::raw(format!("Tile: {:?}", tile_kind))));
-
-    lines.push(Line::from("Tile art:"));
-    if block_art_12x8_lines_for_position(app, pos, &mut lines) {
-        lines.push(Line::from(""));
-    } else {
-        lines.extend(empty_art_12x8_lines_for_position());
-        lines.push(Line::from(""));
-    }
-
-    let content = Paragraph::new(lines)
-        .style(Style::default().fg(Color::White))
-        .alignment(Alignment::Left)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title("Look at")
-                .style(Style::default().fg(Color::White)),
-        );
-    f.render_widget(content, area);
-}
