+++ crates/client/src/ui/panels/mod.rs
@@ -3,6 +3,8 @@ pub mod crafting;
 pub mod credits;
 pub mod help;
 pub mod inventory;
+pub mod look;
+pub mod menu;
 pub mod quit;
 
 pub use body::render_body_panel;
@@ -10,4 +12,6 @@ pub use crafting::render_crafting_panel;
 pub use credits::render_credits_panel;
 pub use help::render_help_panel;
 pub use inventory::{get_player_inventory, render_inventory_list_only, render_inventory_panel};
+pub use look::render_look_panel;
+pub use menu::render_menu_panel;
 pub use quit::render_quit_panel;
